Crate winter_rand_utils

This crates contains functions for generating random values.

These functions are intended to be used in tests, benchmarks, and examples. When compiled to WebAssembly target, all of the functions are omitted.

Functions

prng_array

Returns an array of value of the specified type and the specified length generated pseudo-randomly from the specified seed.

prng_vector

Returns a vector of value of the specified type and the specified length generated pseudo-randomly from the specified seed.

rand_array

Returns an array of random value of the specified type and the specified length.

rand_value

Returns a single random value of the specified type.

rand_vector

Returns a vector of random value of the specified type and the specified length.

shuffle

Randomly shuffles slice elements.